# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/google/callback": {
      "get": {
        "description": "accessToken과 refreshToken은 쿠키로 관리됩니다",
        "operationId": "AuthController_googleAuthRedirect",
        "parameters": [
          {
            "name": "service",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponseDto"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string"
                        },
                        "responseAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "data": {
                          "$ref": "#/components/schemas/MemberPayloadWithoutTypeDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Google Oauth Code로 서비스에 로그인",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "*쿠키에 refreshToken을 설정해야 합니다.",
        "operationId": "AuthController_refreshAccessToken",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponseDto"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string"
                        },
                        "responseAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "data": {
                          "type": "string",
                          "example": "ok"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "refreshToken": []
          }
        ],
        "summary": "Access Token을 Refresh하기",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponseDto"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string"
                        },
                        "responseAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "data": {
                          "$ref": "#/components/schemas/MemberPayloadWithoutTypeDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "내 정보 가져오기",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "TAPIE API",
    "description": "TAPIE System API",
    "version": "0.0.8",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8877",
      "description": "Local Development"
    },
    {
      "url": "https://tapie-api-development.vport.dev",
      "description": "Remote Development"
    },
    {
      "url": "https://api.tapie.kr",
      "description": "Remote Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "accessToken": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "MemberPayloadWithoutTypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "name"
        ]
      },
...
    }
  }
}